Little corrections
The Rpcs cannot receive or send Transform information. Instead, send Vector3 and Quaternion info separately.

Also, the GameManager was not networked, so a GameManagerCreator instantiates it from the masterClient with photono's PhotonNetwork.Instantiate method.

Remember that every single object that uses the PhotonNetwork needs a Photon View component attached to it in order to work correctly.